introduce bundle command
