+++ src/plugin.rs
@@ -1,9 +1,16 @@
+mod action;
 mod async_action;
 mod cli;
 mod handler;
 mod tool;
 
-use crate::plugin::{async_action::*, cli::*, handler::*, tool::*};
+use crate::plugin::{
+    action::ActionPlugin,
+    async_action::{AsyncAction, AsyncActionPlugin},
+    cli::CliPlugin,
+    handler::HandlerPlugin,
+    tool::ToolPlugin,
+};
 use dip::bevy::app::{App, Plugin};
 
 pub struct DipCliPlugin;
@@ -13,7 +20,7 @@ impl Plugin for DipCliPlugin {
         app.add_plugin(CliPlugin::<AsyncAction>::application())
             .add_plugin(ActionPlugin)
             .add_plugin(AsyncActionPlugin)
-            .add_plugin(ToolPlugin)
-            .add_plugin(HandlerPlugin);
+            .add_plugin(HandlerPlugin)
+            .add_plugin(ToolPlugin);
     }
 }
