+++ src/plugin/action/bundle.rs
@@ -1,0 +1,9 @@
+use dip::cli::SubcommandPlugin;
+
+#[derive(SubcommandPlugin, clap::Subcommand, Clone, Debug)]
+pub enum BundleAction {
+    Apply {
+        #[clap(short, long, required = false)]
+        verbose: bool,
+    },
+}
