+++ src/plugin/action/tool.rs
@@ -1,0 +1,7 @@
+use dip::cli::SubcommandPlugin;
+
+#[derive(SubcommandPlugin, clap::Subcommand, Clone, Debug)]
+pub enum ToolAction {
+    List,
+    Add { name: String },
+}
