+++ src/plugin/async_action.rs
@@ -1,5 +1,5 @@
 use crate::{
-    plugin::BuildAction,
+    plugin::action::BuildAction,
     resource::tool::{Tool, ToolResult},
 };
 use dip::{bevy::ecs::event::EventReader, core::task::async_action};
